Мелкие изменения

+++ WpfApp2/Room.cs
@@ -26,7 +26,7 @@ public abstract class Room : Page
     protected bool _isUpKeyPressed = false, _isDownKeyPressed = false,
                    _isLeftKeyPressed = false, _isRightKeyPressed = false;
 
-    protected const float _friction = 0.88F, _speed = 1.5F;
+    protected const double _friction = 0.88F, _speed = 1.5F;
 
     protected bool _isPossibleUpwardMovement, _isPossibleDownwardMovement,
                    _isPossibleLeftwardMovement, _isPossibleRightwardMovement;
